Reconstruct the framework of base_inference
