+++ align_anything/configs/evaluation/deepspeed/ds_z2_offload_config.json
@@ -1,5 +1,5 @@
 {
-  "inference_batch_size": 128,
+  "inference_batch_size": 1,
   "train_batch_size": 128,
   "train_micro_batch_size_per_gpu": 16,
   "gradient_accumulation_steps": null,
