+++ align_anything/evaluation/benchmarks/ARC/ds_infer.py
@@ -18,13 +18,12 @@ import json
 import pickle
 import argparse
 import torch.distributed as dist
-from align_anything.evaluation.inference.base_inference import BaseInferencer_deepspeed
+from align_anything.evaluation.inference.ds_inference import BaseInferencer_deepspeed
 from align_anything.evaluation.dataloader.base_dataloader import BaseDataLoader
 from typing import List, Dict
 from align_anything.utils.tools import read_eval_cfgs, dict_to_namedtuple, update_dict, custom_cfgs_to_dict
 from align_anything.utils.template_registry import get_template_class
 from align_anything.evaluation.data_type import InferenceInput, InferenceOutput
-from align_anything.evaluation.inference.base_inference import update_results
 
 class ARCDataLoader(BaseDataLoader):
     def get_task_names(self):
