+++ align_anything/evaluation/benchmarks/MMVet/ds_infer.py
@@ -15,13 +15,13 @@
 
 import os
 import argparse
-from align_anything.evaluation.inference.base_inference import BaseInferencer_deepspeed, ListDataset
+from align_anything.evaluation.inference.ds_inference import BaseInferencer_deepspeed, ListDataset
 from align_anything.evaluation.dataloader.base_dataloader import BaseDataLoader
 from typing import List, Dict, Any
 from align_anything.utils.tools import read_eval_cfgs, dict_to_namedtuple, update_dict, custom_cfgs_to_dict
 from align_anything.utils.template_registry import get_template_class
 from align_anything.evaluation.data_type import InferenceInput, InferenceOutput
-from align_anything.evaluation.inference.base_inference import get_rank
+from align_anything.evaluation.inference.ds_inference import get_rank
 from torch.nn.utils.rnn import pad_sequence
 import torch.distributed as dist
 from torch.utils.data import DataLoader, DistributedSampler
